Everything is finally fixed
I was calling the wrong set position, which meant the array was drawing
every cube in the same place. Also changed the background colour to
black to appear as if we're in space. Looking to change the sphere to a
Ti-fighter like ship, add change the cubes to asteroid like shaped, add
in collision, powers up and on screen info, and maybe an end goal, like
getting to a distant planet.