Added In Bullets
Created a vector array of bullets, attached a mesh with a texture to a
bullet GameEntity, and every time space pressed the game draws new a
Bullet Game Entity, with a delay between so that there is gap between
bullets.
currently working on collision for the bullets with the asteroids. For
this I'm using AABB collision, where I take the min pos and max pos of
each mesh relative to the position of each mesh, then testing the usual
aabb test to test if a collision has occurred, and then deleting the
asteroid and the bullet, and cout-ing true to the console.