Cleared Up The Project
Got rid of some folders and files that weren't necessary anymore for my
program. I also  changed the player model to a sphere and gave it a
death star type texture. Lastly, I added in collision for the player
with the asteroids, which takes away one of the player's lives (starts
with 3 lives at first) and deletes the asteroid that the player hit.