Added in Text, Restart function, and Finished
I added in a new second screen that appears along side the main game
screen. This second acts as an info screen that gives the user its
score, its remaining lives, and instructions as to how to play the game.
When the user dies, the screen will clear and display a message telling
them that they have lost, it will tell them their final score, and after
a short delay (to avoid immediate action), it will tell the user that it
can restart the game by pressing space. This will reset the needed
variables back to their original value, it will delete the array and
objects, and then call the InitEntities function to re-Initialise the
entities, which, along with a resetted score and lives, will reset the
game for the user.